hu2/codigos de estado de prestamos- mapeo model->entity

+++ domain/model/src/main/java/co/pragma/model/estadosolicitud/gateways/EstadoSolicitudRepository.java
@@ -4,6 +4,6 @@ import co.pragma.model.estadosolicitud.EstadoSolicitud;
 import reactor.core.publisher.Mono;
 
 public interface EstadoSolicitudRepository {
-    Mono<EstadoSolicitud> findById(int id);
+    Mono<EstadoSolicitud> findById(short id);
     Mono<EstadoSolicitud> findByNombre(String nombre);
 }
